fix: update documentation for reset_to_98 function and remove unnecessary comments

+++ pointers_arrays_strings/0-reset_to_98.c
@@ -1,10 +1,11 @@
 #include "main.h"
 
-/*
-
-*/
+/**
+ * reset_to_98 - change the number to 98
+ * @n: is the number changed to 98
+ */
 
 void reset_to_98(int *n)
 {
-	*n = (int)98;
+	*n = 98;
 }
